Initial commit of fantasyGameInventory
ATBS Ch. 5 exercise with dictionaries.  Initial commit includes 'displayInventory' function to display a newline seperated output of any passed inventory value where the data is stored as a dict in the format of {'Item': 'Total'} where the total is an integer.  It also totals the amount of items in the inventory and displays that on the final line of the output.