Initial commit of randomQuizGenerator.py
Working with files and paths in this practical exercise from Automate the Boring stuff, I created a random test generator that generates a given number (35) of tests, each with a random selection of questions, along with multiple choice answers including one (1) possible correct answer for each question along with the answer key for each quiz (35 tests + 35 keys = 70 total files output).  This isn't exactly a direct lift from the ATBS exercise as I didn't want to clutter up my directory with a bunch of rando-text files, so I used the info on Path and os from the chapter to ensure the files go directly to their own quiz directory (which should also be included in this commit).  I also fiddled around with prettying up the test a bit with things like class_name, header_pad, and student_date_line variables. All in all, an interesting exercise.