Initial commit of madlibs.py and the madlibs files
Currently functioning in so far as opening and closing of proper files in the correct locations, prompting user to choose a madlib to play, and prompting users for individual words/numbers etc. for input.  Still have to find a way to reintegrate the user responses back into the text in the right order.