Cleaned up usage check and added file search generator function.
After feedback from groucho on Learn Programming discord, I changed up the usage return by putting it into a sanity check at the head of the file and removing the elif containing the condition check.  Also realized in testing that the script as it was didn't handle ignoring directories at all & errored out when it hit one.  Changed line 88 to call a generator function that yields only filenames and skips directories entirely.